# server/src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Antimony API",
        "contact": {
            "name": "Institute for Networking at OST",
            "url": "https://www.ost.ch/en/research-and-consulting-services/computer-science/ins-institute-for-network-and-security",
            "email": "antimony@network.garden"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/labs": {
            "get": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labs"
                ],
                "summary": "Get all labs",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.OkResponse-array_lab_LabOut"
                        }
                    },
                    "401": {
                        "description": "The user isn't authorized"
                    },
                    "403": {
                        "description": "Access to the resource was denied. Details in the request body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "498": {
                        "description": "The provided access token is not valid"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labs"
                ],
                "summary": "Create a new lab",
                "parameters": [
                    {
                        "description": "The lab",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lab.LabIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The ID of the newly created lab",
                        "schema": {
                            "$ref": "#/definitions/utils.OkResponse-string"
                        }
                    },
                    "401": {
                        "description": "The user isn't authorized"
                    },
                    "403": {
                        "description": "Access to the resource was denied. Details in the request body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "498": {
                        "description": "The provided access token is not valid"
                    }
                }
            }
        },
        "/labs/{id}": {
            "put": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labs"
                ],
                "summary": "Update an existing lab",
                "parameters": [
                    {
                        "description": "The lab with updated values",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/lab.LabIn"
                        }
                    },
                    {
                        "type": "string",
                        "description": "The ID of the lab to edit",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "The user isn't authorized"
                    },
                    "403": {
                        "description": "Access to the resource was denied. Details in the request body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "The request was invalid. Details in the response body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "498": {
                        "description": "The provided access token is not valid"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BasicAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "labs"
                ],
                "summary": "Delete an existing lab",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The ID of the lab to delete",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "The user isn't authorized"
                    },
                    "403": {
                        "description": "Access to the resource was denied. Details in the request body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "The request was invalid. Details in the response body.",
                        "schema": {
                            "$ref": "#/definitions/utils.ErrorResponse"
                        }
                    },
                    "498": {
                        "description": "The provided access token is not valid"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "deployment.NodeState": {
            "type": "string",
            "enum": [
                "running"
            ],
            "x-enum-varnames": [
                "running"
            ]
        },
...
        "lab.InstanceNode": {
            "type": "object",
            "properties": {
                "containerId": {
                    "type": "string"
                },
                "containerName": {
                    "type": "string"
                },
                "ipv4": {
                    "type": "string"
                },
                "ipv6": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "port": {
                    "type": "integer"
                },
                "state": {
                    "$ref": "#/definitions/deployment.NodeState"
                },
                "user": {
                    "type": "string"
                },
                "webSSH": {
                    "type": "string"
                }
            }
        },
        "lab.InstanceOut": {
            "type": "object",
            "properties": {
                "deployed": {
                    "type": "string"
                },
                "edgesharkLink": {
                    "type": "string"
                },
                "latestStateChange": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lab.InstanceNode"
                    }
                },
                "recovered": {
                    "type": "boolean"
                },
                "state": {
                    "$ref": "#/definitions/lab.InstanceState"
                }
            }
        },
        "lab.InstanceState": {
            "type": "integer",
            "enum": [
                0,
                1,
                2,
                3,
                -1,
                -2
            ],
            "x-enum-varnames": [
                "deploying",
                "running",
                "stopping",
                "failed",
                "inactive",
                "scheduled"
            ]
        },
        "lab.LabIn": {
            "type": "object",
            "required": [
                "topologyId"
            ],
            "properties": {
                "endTime": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "topologyId": {
                    "type": "string"
                }
            }
        },
        "lab.LabOut": {
            "type": "object",
            "properties": {
                "collectionId": {
                    "type": "string"
                },
                "creator": {
                    "$ref": "#/definitions/user.UserOut"
                },
                "endTime": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "instance": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/lab.InstanceOut"
                        }
                    ],
                    "x-nullable": true
                },
                "instanceName": {
                    "type": "string",
                    "x-nullable": true
                },
                "name": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                },
                "topologyId": {
                    "type": "string"
                }
            }
        },
...
        "user.UserOut": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "utils.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "utils.OkResponse-array_lab_LabOut": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/lab.LabOut"
                    }
                }
            }
        },
...
        "utils.OkResponse-string": {
            "type": "object",
            "properties": {
                "payload": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        }
    }
}